Update download.py
Hey Samuel, 

first of all, thank you very much for your code! I've reworked it a bit and added a few features, such as a GUI, batch processing, and a more robust method for merging PDFs. If you like it, feel free to adopt it for your project so others can benefit from it :-)

Mirko